usageStatsReport uses StatsSummary model
